requirements.txt
biopython
requests
tqdm
pydantic
pydantic-ai
openai
transformers
accelerate
pandas